Convert value to bigint before passing to viem (#8)
Without converting to a BigInt, this will fail with custom `value` parameters as they are often converted to hexadecimal format before being passed to `sendTransaction`. In such a case, viem will not properly compute the value, and it will throw an "Insufficient funds" error claiming that the account does not have enough funds to cover `gas * price + value`.